requirements.txt
gunicorn
lxml
mollie-api-python
requests
Faker
beautifulsoup4